Downloads/metabit-web-app-kyc-main (12)/metabit-web-app-kyc-main/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-telegram-bot==20.7
requests==2.31.0
python-dotenv==1.0.0
cloudinary==1.36.0
psycopg2-binary==2.9.9
httpx~=0.25.2
Flask==2.2.2
Flask-SQLAlchemy==3.0.2
Flask-Cors==3.0.10
werkzeug==2.2.3
aiohttp==3.9.1
psutil==5.9.6